+++ crates/jiff/src/civil/weekday.rs
@@ -82,6 +82,7 @@ use crate::error::Error;
 /// ```
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
+#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 #[repr(u8)]
 #[allow(missing_docs)]
 pub enum Weekday {